requirements.txt
argparse
openai==0.27.8
Pillow
pyfiglet
requests
tk